refactor(desktop): make the chrome bar's actions slot the only way out
The report header in a pane spread its own children, so its actions only
looked right-aligned while the title was long enough to push them: on a
short title they floated mid-bar, 250px short of the edge. Every caller
now passes trailing controls through ChromeBar's actions slot, which is
the one thing that reaches the far end, and the last hand-written bar
(the freeform canvas toolbar) moves onto the component too.

DetailBackLink was two components in one: a crumb trail when given a
report, a back arrow otherwise, and its main caller passed a `to` and a
`label` that the crumb branch ignored. The trail is now
ReportBreadcrumbs, and the detail frame no longer takes a back link it
does not draw.

Deletes the rail's second, unused answer to "which destination is this
href", which still read the route pattern the branch stopped trusting.
